Externalizing Table component
